+++ L1Trigger/L1THGCal/plugins/fe_codecs/HGCalTriggerCellThresholdCodec.cc
@@ -79,7 +79,7 @@ setDataPayloadImpl(const l1t::HGCFETriggerDigi& digi)
     conf.addParameter<uint32_t>   ("tdcnBits",      codecImpl_.tdcnBits());
     conf.addParameter<double>     ("tdcOnsetfC",    codecImpl_.tdcOnsetfC());
     conf.addParameter<uint32_t>   ("triggerCellTruncationBits", codecImpl_.triggerCellTruncationBits());
-    conf.addParameter<float>        ("TCThreshold_fC", codecImpl_.TCThreshold_fC());
+    conf.addParameter<double>        ("TCThreshold_fC", codecImpl_.TCThreshold_fC());
     HGCalTriggerCellThresholdCodec codecInput(conf);
     codecInput.setGeometry(geometry_);
     digi.decode(codecInput,data_);
